Add files via upload
This is an analysis using R programming for a bike-sharing company. The marketing manager wanted to develop new marketing strategies to convert customers into annual members. To do this, however, they need to understand through visualization how the customers and annual members use the company's bikes differently.  This will then be used to derive insights for the business task.